refactor: 50% speedup in both parts
This is a weird one.
Doing the extra test for both is 10% slower for part 1 and 15% faster for part2.
But somehow, splitting the is_possible_* checks into two functions (one for part 1 without the early check, one for part 2 inncluding it) speeds up both part 1 and part 2 by 50% overall.
Maybe it allows some extra inlining or similar?